Create README.md
# Attendance-Management-Application-Student-version-using-Google-Firebase-database
This is Student version of Online Attendance Management application project .Basically this project has two app one is administrative app 
and other is student app. So this app is Student Application. This whole project is distributed in app as well website as per need and 
both of then is connect with a comman database that works same as the app does.